split into sway & Hyprland
