+++ nwg_shell_config/shell/settings-hyprland
@@ -1,0 +1,92 @@
+{
+  "keyboard-layout": "us",
+  "autotiling-workspaces": "",
+  "autotiling-on": true,
+  "autotiling-limit": false,
+  "autotiling-output-limits": {},
+  "appindicator": true,
+  "night-lat": -1,
+  "night-long": -1,
+  "night-temp-low": 4500,
+  "night-temp-high": 6500,
+  "night-gamma": 1.0,
+  "night-on": true,
+  "terminal": "",
+  "file-manager": "",
+  "editor": "",
+  "browser": "",
+  "panel-preset": "preset-0",
+  "panel-custom": "",
+  "show-on-startup": true,
+  "keyboard-use-settings": true,
+  "keyboard-identifier": "",
+  "keyboard-xkb-layout": "us",
+  "keyboard-xkb-variant": "",
+  "keyboard-repeat-delay": 300,
+  "keyboard-repeat-rate": 40,
+  "keyboard-xkb-capslock": "disabled",
+  "keyboard-xkb-numlock": "disabled",
+  "keyboard-custom-name": "",
+  "keyboard-custom-value": "",
+  "pointer-use-settings": true,
+  "pointer-identifier": "",
+  "pointer-accel-profile": "flat",
+  "pointer-pointer-accel": 0.0,
+  "pointer-natural-scroll": "disabled",
+  "pointer-scroll-factor": 1.0,
+  "pointer-left-handed": "disabled",
+  "pointer-custom-name": "",
+  "pointer-custom-value": "",
+  "touchpad-use-settings": true,
+  "touchpad-identifier": "",
+  "touchpad-accel-profile": "flat",
+  "touchpad-pointer-accel": 0.0,
+  "touchpad-natural-scroll": "disabled",
+  "touchpad-scroll-factor": 1.0,
+  "touchpad-scroll-method": "two_finger",
+  "touchpad-left-handed": "disabled",
+  "touchpad-tap": "enabled",
+  "touchpad-tap-button-map": "lrm",
+  "touchpad-drag": "enabled",
+  "touchpad-drag-lock": "disabled",
+  "touchpad-dwt": "enabled",
+  "touchpad-middle-emulation": "enabled",
+  "touchpad-custom-name": "",
+  "touchpad-custom-value": "",
+  "lockscreen-use-settings": true,
+  "lockscreen-locker": "gtklock",
+  "lockscreen-background-source": "local",
+  "lockscreen-custom-cmd": "",
+  "lockscreen-timeout": 1200,
+  "sleep-cmd": "swaymsg \"output * dpms off\"",
+  "sleep-timeout": 1800,
+  "resume-cmd": "swaymsg \"output * dpms on\"",
+  "before-sleep": "",
+  "backgrounds-custom-path": "",
+  "backgrounds-use-custom-path": false,
+  "background-dirs": [
+    "/usr/share/backgrounds/nwg-shell"
+  ],
+  "background-dirs-once-set": false,
+  "unsplash-width": 1920,
+  "unsplash-height": 1080,
+  "unsplash-keywords": [
+    "nature",
+    "water",
+    "landscape"
+  ],
+  "help-font-size": 14,
+  "help-layer-shell": true,
+  "help-keyboard": false,
+  "gtklock-disable-input-inhibitor": false,
+  "gtklock-idle-timeout": 10,
+  "gtklock-logout-command": "swaymsg exit",
+  "gtklock-playerctl": false,
+  "gtklock-powerbar": false,
+  "gtklock-poweroff-command": "systemctl -i poweroff",
+  "gtklock-reboot-command": "systemctl reboot",
+  "gtklock-suspend-command": "systemctl suspend",
+  "gtklock-time-format": "%H:%M:%S",
+  "gtklock-userinfo": false,
+  "gtklock-userswitch-command": ""
+}
