+++ setup.py
@@ -9,7 +9,7 @@ def read(f_name):
 
 setup(
     name='nwg-shell-config',
-    version='0.4.21',
+    version='0.4.22',
     description='nwg-shell configuration utility',
     packages=find_packages(),
     include_package_data=True,
@@ -25,6 +25,8 @@ setup(
     entry_points={
         'gui_scripts': [
             'nwg-shell-config = nwg_shell_config.main:main',
+            'nwg-shell-config-sway = nwg_shell_config.main-sway:main',
+            'nwg-shell-config-hyprland = nwg_shell_config.main-hyprland:main',
             'nwg-lock = nwg_shell_config.locker:main',
             'nwg-shell-help = nwg_shell_config.help:main',
             'nwg-autotiling = nwg_shell_config.autotiling:main',
